+++ src/main/java/plugily/projects/murdermystery/arena/ArenaManager.java
@@ -236,8 +236,7 @@ public class ArenaManager {
     Debugger.debug("[{0}] Initial leave attempt for {1}", arena.getId(), player.getName());
     long start = System.currentTimeMillis();
 
-    MMGameLeaveAttemptEvent event = new MMGameLeaveAttemptEvent(player, arena);
-    Bukkit.getPluginManager().callEvent(event);
+    Bukkit.getPluginManager().callEvent(new MMGameLeaveAttemptEvent(player, arena));
     User user = plugin.getUserManager().getUser(player);
     if(user.getStat(StatsStorage.StatisticType.LOCAL_SCORE) > user.getStat(StatsStorage.StatisticType.HIGHEST_SCORE)) {
       user.setStat(StatsStorage.StatisticType.HIGHEST_SCORE, user.getStat(StatsStorage.StatisticType.LOCAL_SCORE));
@@ -265,13 +264,15 @@ public class ArenaManager {
     }
 
     arena.getScoreboardManager().removeScoreboard(user);
-    if(Role.isRole(Role.MURDERER, player)) {
+
+    boolean playerHasMurdererRole = Role.isRole(Role.MURDERER, player);
+    if(playerHasMurdererRole) {
       arena.removeFromMurdererList(player);
     }
-    //-1 cause we didn't remove player yet
     if(arena.getArenaState() == ArenaState.IN_GAME && !user.isSpectator()) {
+      //-1 cause we didn't remove player yet
       if(arena.getPlayersLeft().size() - 1 > 1) {
-        if(Role.isRole(Role.MURDERER, player)) {
+        if(playerHasMurdererRole) {
           if(arena.getMurdererList().isEmpty()) {
             List<Player> players = new ArrayList<>();
             for(Player gamePlayer : arena.getPlayersLeft()) {
@@ -373,8 +374,7 @@ public class ArenaManager {
     Debugger.debug("[{0}] Stop game event initialized with quickStop {1}", arena.getId(), quickStop);
     long start = System.currentTimeMillis();
 
-    MMGameStopEvent gameStopEvent = new MMGameStopEvent(arena);
-    Bukkit.getPluginManager().callEvent(gameStopEvent);
+    Bukkit.getPluginManager().callEvent(new MMGameStopEvent(arena));
     arena.setArenaState(ArenaState.ENDING);
     if(quickStop) {
       arena.setTimer(2);
@@ -484,7 +484,7 @@ public class ArenaManager {
 
     formatted = StringUtils.replace(formatted, "%murderer%", (arena.lastAliveMurderer() ? "" : ChatColor.STRIKETHROUGH) + murders.toString());
 
-    formatted = StringUtils.replace(formatted, "%murderer_kills%", String.valueOf(murdererKills));
+    formatted = StringUtils.replace(formatted, "%murderer_kills%", Integer.toString(murdererKills));
     formatted = StringUtils.replace(formatted, "%hero%", arena.isCharacterSet(Arena.CharacterType.HERO)
         ? arena.getCharacter(Arena.CharacterType.HERO).getName() : chatManager.colorMessage("In-Game.Messages.Game-End-Messages.Winners.Nobody"));