+++ src/main/java/plugily/projects/murdermystery/handlers/rewards/RewardsFactory.java
@@ -101,7 +101,7 @@ public class RewardsFactory {
     String formatted = command;
     formatted = StringUtils.replace(formatted, "%ARENA-ID%", arena.getId());
     formatted = StringUtils.replace(formatted, "%MAPNAME%", arena.getMapName());
-    formatted = StringUtils.replace(formatted, "%PLAYERAMOUNT%", String.valueOf(arena.getPlayers().size()));
+    formatted = StringUtils.replace(formatted, "%PLAYERAMOUNT%", Integer.toString(arena.getPlayers().size()));
     return formatted;
   }