+++ src/main/java/plugily/projects/murdermystery/handlers/sign/SignManager.java
@@ -83,12 +83,13 @@ public class SignManager implements Listener {
       || !ComplementAccessor.getComplement().getLine(e, 0).equalsIgnoreCase("[murdermystery]")) {
       return;
     }
-    if(ComplementAccessor.getComplement().getLine(e, 1).isEmpty()) {
+    String line1 = ComplementAccessor.getComplement().getLine(e, 1);
+    if(line1.isEmpty()) {
       e.getPlayer().sendMessage(chatManager.getPrefix() + chatManager.colorMessage("Signs.Please-Type-Arena-Name"));
       return;
     }
     for(Arena arena : ArenaRegistry.getArenas()) {
-      if(!arena.getId().equalsIgnoreCase(ComplementAccessor.getComplement().getLine(e, 1))) {
+      if(!arena.getId().equalsIgnoreCase(line1)) {
         continue;
       }
       for(int i = 0; i < signLines.size(); i++) {
@@ -110,13 +111,15 @@ public class SignManager implements Listener {
   private String formatSign(String msg, Arena a) {
     String formatted = msg;
     formatted = StringUtils.replace(formatted, "%mapname%", a.getMapName());
-    if(a.getPlayers().size() >= a.getMaximumPlayers()) {
+
+    int maxPlayers = a.getMaximumPlayers();
+    if(a.getPlayers().size() >= maxPlayers) {
       formatted = StringUtils.replace(formatted, "%state%", chatManager.colorMessage("Signs.Game-States.Full-Game"));
     } else {
       formatted = StringUtils.replace(formatted, "%state%", gameStateToString.get(a.getArenaState()));
     }
-    formatted = StringUtils.replace(formatted, "%playersize%", String.valueOf(a.getPlayers().size()));
-    formatted = StringUtils.replace(formatted, "%maxplayers%", String.valueOf(a.getMaximumPlayers()));
+    formatted = StringUtils.replace(formatted, "%playersize%", Integer.toString(a.getPlayers().size()));
+    formatted = StringUtils.replace(formatted, "%maxplayers%", Integer.toString(maxPlayers));
     formatted = chatManager.colorRawMessage(formatted);
     return formatted;
   }