+++ src/main/java/plugily/projects/murdermystery/user/UserManager.java
@@ -59,12 +59,12 @@ public class UserManager {
 
   public User getUser(Player player) {
     for(User user : users) {
-      if(user.getPlayer().equals(player)) {
+      if(user.getUniqueId().equals(player.getUniqueId())) {
         return user;
       }
     }
     Debugger.debug("Registering new user {0} ({1})", player.getUniqueId(), player.getName());
-    User user = new User(player);
+    User user = new User(player.getUniqueId());
     users.add(user);
     return user;
   }