+++ src/main/java/plugily/projects/murdermystery/user/data/FileStats.java
@@ -41,7 +41,7 @@ public class FileStats implements UserDatabase {
 
   @Override
   public void saveStatistic(User user, StatsStorage.StatisticType stat) {
-    config.set(user.getPlayer().getUniqueId().toString() + "." + stat.getName(), user.getStat(stat));
+    config.set(user.getUniqueId().toString() + "." + stat.getName(), user.getStat(stat));
     ConfigUtils.saveConfig(plugin, config, "stats");
   }
 
@@ -49,7 +49,7 @@ public class FileStats implements UserDatabase {
   public void saveAllStatistic(User user) {
     for(StatsStorage.StatisticType stat : StatsStorage.StatisticType.values()) {
       if(!stat.isPersistent()) continue;
-      config.set(user.getPlayer().getUniqueId().toString() + "." + stat.getName(), user.getStat(stat));
+      config.set(user.getUniqueId().toString() + "." + stat.getName(), user.getStat(stat));
     }
     ConfigUtils.saveConfig(plugin, config, "stats");
   }
@@ -57,7 +57,7 @@ public class FileStats implements UserDatabase {
   @Override
   public void loadStatistics(User user) {
     for(StatsStorage.StatisticType stat : StatsStorage.StatisticType.values()) {
-      user.setStat(stat, config.getInt(user.getPlayer().getUniqueId().toString() + "." + stat.getName(), 0));
+      user.setStat(stat, config.getInt(user.getUniqueId().toString() + "." + stat.getName(), 0));
     }
   }